+++ pkg/config/configyaml.go
@@ -40,29 +40,46 @@ type ConfigYGlobal struct {
 	Locale             string
 }
 type ConfigYConnection struct {
-	QueueManager string `yaml:"queueManager"`
-	User         string
-	Client       string `yaml:"clientConnection" default:"false"`
-	Password     string
-	ReplyQueue   string `yaml:"replyQueue"`
-	ReplyQueue2  string `yaml:"replyQueue2"`
-	CcdtUrl      string `yaml:"ccdtUrl"`
-	ConnName     string `yaml:"connName"`
-	Channel      string `yaml:"channel"`
+	QueueManager     string `yaml:"queueManager"`
+	User             string
+	Client           string `yaml:"clientConnection" default:"false"`
+	Password         string
+	ReplyQueue       string `yaml:"replyQueue" `
+	ReplyQueue2      string `yaml:"replyQueue2"`
+	DurableSubPrefix string `yaml:"durableSubPrefix"`
+	CcdtUrl          string `yaml:"ccdtUrl"`
+	ConnName         string `yaml:"connName"`
+	Channel          string `yaml:"channel"`
 }
 type ConfigYObjects struct {
-	Queues                    []string
+	Queues        []string
+	Channels      []string
+	Topics        []string
+	Subscriptions []string
+	// These are left here for now so they can be recognised but will cause an error because the
+	// real values have moved
 	QueueSubscriptionSelector []string `yaml:"queueSubscriptionSelector"`
-	Channels                  []string
-	Topics                    []string
-	Subscriptions             []string
-	ShowInactiveChannels      string `yaml:"showInactiveChannels" default:"false"`
+	ShowInactiveChannels      string   `yaml:"showInactiveChannels" default:"false"`
 }
+
+type ConfigYFilters struct {
+	HideSvrConnJobname        string   `yaml:"hideSvrConnJobname" default:"false"`
+	ShowInactiveChannels      string   `yaml:"showInactiveChannels" default:"false"`
+	QueueSubscriptionSelector []string `yaml:"queueSubscriptionSelector"`
+}
+
+type ConfigMoved struct {
+	QueueSubscriptionSelector string
+	ShowInactiveChannels      string
+}
+
+var cfMoved ConfigMoved
 
 func ReadConfigFile(f string, cmy interface{}) error {
 
 	data, e2 := ioutil.ReadFile(f)
 	if e2 == nil {
+		// fmt.Printf("Unparsed Data is\n %s\n", string(data))
 		e2 = yaml.Unmarshal(data, cmy)
 	}
 
@@ -70,7 +87,7 @@ func ReadConfigFile(f string, cmy interface{}) error {
 }
 
 // The Go YAML parsing is not what you might expect for booleans - you are
-// apparently unable to set a default of "true" for missing fields. So we read ut
+// apparently unable to set a default of "true" for missing fields. So we read it
 // as a string and parse that. The caller also sends in the default value if the string
 // cannot be decoded.
 func asBool(s string, def bool) bool {
@@ -84,11 +101,15 @@ func asBool(s string, def bool) bool {
 
 // This handles the configuration parameters that are common to all the collectors. The individual
 // collectors call similar code for their own specific attributes
-func CopyYamlConfig(cm *Config, cyg ConfigYGlobal, cyc ConfigYConnection, cyo ConfigYObjects) {
+func CopyYamlConfig(cm *Config, cyg ConfigYGlobal, cyc ConfigYConnection, cyo ConfigYObjects, cyf ConfigYFilters) {
+
 	cm.CC.UseStatus = CopyParmIfNotSetBool("global", "useObjectStatus", asBool(cyg.UseObjectStatus, true))
 	cm.CC.UseResetQStats = CopyParmIfNotSetBool("global", "useResetQStats", asBool(cyg.UseResetQStats, false))
 	cm.CC.UsePublications = CopyParmIfNotSetBool("global", "usePublications", asBool(cyg.UsePublications, true))
-	cm.CC.ShowInactiveChannels = CopyParmIfNotSetBool("objects", "showInactiveChannels", asBool(cyo.ShowInactiveChannels, false))
+
+	cm.CC.ShowInactiveChannels = CopyParmIfNotSetBool("filters", "showInactiveChannels", asBool(cyf.ShowInactiveChannels, false))
+	cm.CC.HideSvrConnJobname = CopyParmIfNotSetBool("filters", "hideSvrConnJobname", asBool(cyf.HideSvrConnJobname, false))
+	cm.QueueSubscriptionSelector = CopyParmIfNotSetStrArray("filters", "queueSubscriptionSelector", cyf.QueueSubscriptionSelector)
 
 	cm.LogLevel = CopyParmIfNotSetStr("global", "logLevel", cyg.LogLevel)
 	cm.MetaPrefix = CopyParmIfNotSetStr("global", "metaprefix", cyg.MetaPrefix)
@@ -104,14 +125,22 @@ func CopyYamlConfig(cm *Config, cyg ConfigYGlobal, cyc ConfigYConnection, cyo Co
 	cm.CC.ClientMode = CopyParmIfNotSetBool("connection", "clientConnection", asBool(cyc.Client, false))
 	cm.CC.UserId = CopyParmIfNotSetStr("connection", "user", cyc.User)
 	cm.CC.Password = CopyParmIfNotSetStr("connection", "password", cyc.Password)
-	cm.ReplyQ = CopyParmIfNotSetStr("connection", "replyQueue", cyc.ReplyQueue)
+
+	// This is one where we want to use the default non-null value if it's not been provided
+	tmpQ := CopyParmIfNotSetStr("connection", "replyQueue", cyc.ReplyQueue)
+	if tmpQ != "" {
+		cm.ReplyQ = tmpQ
+	}
 	cm.ReplyQ2 = CopyParmIfNotSetStr("connection", "replyQueue2", cyc.ReplyQueue2)
+	cm.CC.DurableSubPrefix = CopyParmIfNotSetStr("connection", "durableSubPrefix", cyc.DurableSubPrefix)
 
 	cm.MonitoredQueues = CopyParmIfNotSetStrArray("objects", "queues", cyo.Queues)
 	cm.MonitoredChannels = CopyParmIfNotSetStrArray("objects", "channels", cyo.Channels)
 	cm.MonitoredTopics = CopyParmIfNotSetStrArray("objects", "topics", cyo.Topics)
 	cm.MonitoredSubscriptions = CopyParmIfNotSetStrArray("objects", "subscriptions", cyo.Subscriptions)
-	cm.QueueSubscriptionSelector = CopyParmIfNotSetStrArray("objects", "queueSubscriptionSelector", cyo.QueueSubscriptionSelector)
+
+	cfMoved.QueueSubscriptionSelector = CopyDeprecatedParmIfNotSetStrArray("objects", "queueSubscriptionSelector", cyo.QueueSubscriptionSelector)
+	cfMoved.ShowInactiveChannels = CopyDeprecatedParmIfNotSetStr("objects", "showInactiveChannels", cyo.ShowInactiveChannels)
 
 	return
 }
@@ -119,7 +148,15 @@ func CopyYamlConfig(cm *Config, cyg ConfigYGlobal, cyc ConfigYConnection, cyo Co
 // If the parameter has already been set by env var or cli, then the value in the main config structure returned. Otherwise
 // the value passed as the "val" parameter - from the YAML version of the configuration elements - is returned
 func CopyParmIfNotSetBool(section string, name string, val bool) bool {
-	v, s := copyParmIfNotSet(section, name)
+	v, s := copyParmIfNotSet(section, name, false)
+	if s {
+		return *(v).(*bool)
+	} else {
+		return val
+	}
+}
+func CopyDeprecatedParmIfNotSetBool(section string, name string, val bool) bool {
+	v, s := copyParmIfNotSet(section, name, true)
 	if s {
 		return *(v).(*bool)
 	} else {
@@ -128,7 +165,16 @@ func CopyParmIfNotSetBool(section string, name string, val bool) bool {
 }
 
 func CopyParmIfNotSetStr(section string, name string, val string) string {
-	v, s := copyParmIfNotSet(section, name)
+	v, s := copyParmIfNotSet(section, name, false)
+	if s {
+		return *(v).(*string)
+	} else {
+		return val
+	}
+}
+
+func CopyDeprecatedParmIfNotSetStr(section string, name string, val string) string {
+	v, s := copyParmIfNotSet(section, name, true)
 	if s {
 		return *(v).(*string)
 	} else {
@@ -137,7 +183,25 @@ func CopyParmIfNotSetStr(section string, name string, val string) string {
 }
 
 func CopyParmIfNotSetStrArray(section string, name string, val []string) string {
-	v, s := copyParmIfNotSet(section, name)
+	v, s := copyParmIfNotSet(section, name, false)
+	if s {
+		return *(v).(*string)
+	} else {
+		// Convert YAML arrays into the single string expected by the mqmetric package
+		s := ""
+		for i := 0; i < len(val); i++ {
+			if i == 0 {
+				s = val[0]
+			} else {
+				s += "," + val[i]
+			}
+		}
+		return s
+	}
+}
+
+func CopyDeprecatedParmIfNotSetStrArray(section string, name string, val []string) string {
+	v, s := copyParmIfNotSet(section, name, true)
 	if s {
 		return *(v).(*string)
 	} else {
@@ -155,7 +219,7 @@ func CopyParmIfNotSetStrArray(section string, name string, val []string) string 
 }
 
 func CopyParmIfNotSetInt(section string, name string, val int) int {
-	v, s := copyParmIfNotSet(section, name)
+	v, s := copyParmIfNotSet(section, name, false)
 	if s {
 		return *(v).(*int)
 	} else {
@@ -167,18 +231,20 @@ func CopyParmIfNotSetInt(section string, name string, val int) int {
 // set by the user via CLI or environment variable
 //
 // Debug of this is handled by direct Printfs as it's run before the logger is configured
-func copyParmIfNotSet(section string, name string) (interface{}, bool) {
+func copyParmIfNotSet(section string, name string, deprecated bool) (interface{}, bool) {
 	k := envVarKey(section, name)
 	if p, ok := configParms[k]; ok {
 		if p.userSet {
-			//fmt.Printf("Returning data from %v\n",p)
+			//fmt.Printf("Returning data from %v\n", p)
 			return p.loc, true
 		} else {
-			//fmt.Printf("Key %s has not been set by user\n",k)
+			//fmt.Printf("Key %s has not been set by user\n", k)
 		}
 	} else {
 		// If this happens, it indicates a problem in one of the config.go files so we leave it in.
-		fmt.Printf("Key %s not found in parms map\n", k)
+		if !deprecated {
+			fmt.Printf("Key %s not found in parms map\n", k)
+		}
 	}
 	return nil, false
 }